requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.2
yt-dlp==2023.12.30
pydub==0.25.1
cachetools==5.3.2
tenacity==8.2.3
